requirements.txt
# Google Cloud Speech-to-Text用
google-cloud-speech>=2.0.0
google-cloud-storage>=2.0.0

# Amazon Transcribe用
boto3>=1.28.0

# 環境変数管理用
python-dotenv>=1.0.0

# 音声ファイル処理用
pydub>=0.25.1
